+++ services/platform/server/web-client/src/pages/DefaultLayout.css
